+++ express/api_routes.js
@@ -1,7 +1,8 @@
 'use strict';
-const { extension } = require('./extension');
+const { extension } = require('../lib/extension');
 const express = require('express');
-const { sessionMiddleware, partnerSessionMiddleware } = require('./middleware/session_middleware');
+const { sessionMiddleware, partnerSessionMiddleware } = require('../lib/middleware/session_middleware');
+const { getApplicationConfig, getUserData } = require('../lib/utils');
 const { ApplicationConfig, ApplicationClient } = require("@gofynd/fdk-client-javascript");
 
 
@@ -12,20 +13,12 @@ function setupProxyRoutes(configData) {
 
     applicationProxyRoutes.use(async (req, res, next) => {
         try {
-            if (req.headers["x-user-data"]) {
-                req.user = JSON.parse(req.headers["x-user-data"]);
-                req.user.user_id = req.user._id;
-            }
-            if (req.headers["x-application-data"]) {
-                req.application = JSON.parse(req.headers["x-application-data"]);
-                req.applicationConfig = new ApplicationConfig({
-                    applicationID: req.application._id,
-                    applicationToken: req.application.token,
-                    domain: extension.cluster,
-                    logLevel: configData.debug ===  true? "debug": null
-                });
-                req.applicationClient = new ApplicationClient(req.applicationConfig);
-            }
+            const user = await getUserData(req.headers["x-user-data"]);
+            const { application, applicationConfig, applicationClient } = await getApplicationConfig(req.headers["x-application-data"], extension)
+            req.user = user;
+            req.application = application;
+            req.applicationConfig = applicationConfig;
+            req.applicationClient = applicationClient;
             next();
         } catch (error) {
             next(error);