+++ spec/tests/webhooksv1.spec.js
@@ -2,9 +2,9 @@
 
 const fdkHelper = require("../helpers/fdk");
 const { clearData } = require("../helpers/setup_db");
-const request = require("../helpers/server")();
+const request = require("../helpers/server");
 const axiosMock = require("./../mocks/axiosv1.mock.js");
-const { SESSION_COOKIE_NAME } = require("../../express/constants");
+const { SESSION_COOKIE_NAME } = require("../../lib/constants");
 const hmacSHA256 = require("crypto-js/hmac-sha256");
 
 function getSignature(body) {
@@ -14,10 +14,13 @@ function getSignature(body) {
 describe("Webhook Integrations for v1", () => {
     let webhookConfig = null;
     let cookie = "";
-    beforeEach(async () => {
+    let fdk_instance = null;
+    beforeAll(async () => {
+        jasmine.DEFAULT_TIMEOUT_INTERVAL = 99999
         webhookConfig = {
-            api_path: '/v1/webhooks',
-            notification_email: 'test@abc.com',
+            api_path: '/v1/webhooksv1',
+            notification_email: 'testmanish@abc.com',
+            subscribed_saleschannel: 'specific',
             event_map: {
                 'company/product/create': {
                     version: '1',
@@ -26,23 +29,24 @@ describe("Webhook Integrations for v1", () => {
                 },
                 'application/coupon/create': {
                     version: '1',
-                    handler: function () { throw Error('test error') }
+                    handler: function () { throw Error('test error3') },
+                    provider: 'rest'
                 }
             }
         }
-        this.fdk_instance = await fdkHelper({
+        fdk_instance = await fdkHelper.initializeFDK({
             access_mode: "offline",
+            debug: true,
             webhook_config: webhookConfig
         });
-        request.app.restApp.use(this.fdk_instance.fdkHandler);
-        request.app.restApp.post('/v1/webhooks', async (req, res, next)=>{
+        request.app.restApp.use(fdk_instance.fdkHandler);
+        request.app.restApp.post('/v1/webhooksv1', async (req, res, next)=>{
             let status = 404;
             try {
-                await this.fdk_instance.webhookRegistry.processWebhook(req);
+                await fdk_instance.webhookRegistry.processWebhook(req);
                 status = 200;
             }
             catch(err) {
-                console.log(err);
                 status = 500;
             }
             return res.status(status).json({"success": status === 200});
@@ -61,9 +65,9 @@ describe("Webhook Integrations for v1", () => {
         expect(response.status).toBe(302);
     });
 
-    afterEach(async () => {
+    afterAll(async () => {
         await clearData();
-        this.fdk_instance.extension._isInitialized = false;
+        fdk_instance.extension._isInitialized = false;
     });
 
     afterAll(() => {
@@ -74,7 +78,7 @@ describe("Webhook Integrations for v1", () => {
     it("Register webhooks", async () => {
         const reqBody = { "company_id": 1, "payload": { "test": true }, "event": {"name": "product", "type": "create", "category": "company", "version": '1'} };
         const res = await request
-            .post(`/v1/webhooks`)
+            .post(`/v1/webhooksv1`)
             .set('cookie', `${SESSION_COOKIE_NAME}_1=${cookie}`)
             .set('x-fp-signature', getSignature(reqBody))
             .send(reqBody);
@@ -85,7 +89,7 @@ describe("Webhook Integrations for v1", () => {
     it("Invalid webhook path", async () => {
         const reqBody = { "company_id": 1, "payload": { "test": true }, "event": {"name": "coupon", "type": "update", "category": "application", "version": '1'} };
         const res = await request
-            .post(`/v1/webhooks`)
+            .post(`/v1/webhooksv1`)
             .set('cookie', `${SESSION_COOKIE_NAME}_1=${cookie}`)
             .set('x-fp-signature', getSignature(reqBody))
             .send(reqBody);
@@ -95,7 +99,7 @@ describe("Webhook Integrations for v1", () => {
     it("Failed webhook handler execution", async () => {
         const reqBody = { "company_id": 1, "payload": { "test": true }, "event": {"name": "coupon", "type": "create", "category": "application", "version": '1'} };
         const res = await request
-            .post(`/v1/webhooks`)
+            .post(`/v1/webhooksv1`)
             .set('cookie', `${SESSION_COOKIE_NAME}_1=${cookie}`)
             .set('x-fp-signature', getSignature(reqBody))
             .send(reqBody);
@@ -106,20 +110,27 @@ describe("Webhook Integrations for v1", () => {
     it("Sync webhooks: Add new", async () => {
         const reqBody = { "company_id": 1, "payload": { "test": true }, "event": {"name": "coupon", "type": "create", "category": "application", "version": '1'} };
         const newMap = {
-            api_path: '/v1/webhooks',
+            api_path: '/v1/webhooksv1',
             notification_email: 'test@abc.com',
             event_map: {
+                'company/product/create': {
+                    version: '1',
+                    handler: function () { },
+                    provider: 'rest'
+                },
                 'application/coupon/create': {
                     version: '1',
-                    handler: function () { }
+                    handler: function () { 
+                        return true;
+                    }
                 }
             }
         }
         const handlerFn = spyOn(newMap.event_map['application/coupon/create'], 'handler');
-        const platformClient = await this.fdk_instance.getPlatformClient('1');
-        await this.fdk_instance.webhookRegistry.syncEvents(platformClient, newMap);
+        const platformClient = await fdk_instance.getPlatformClient('1');
+        await fdk_instance.webhookRegistry.syncEvents(platformClient, newMap);
         const res = await request
-            .post(`/v1/webhooks`)
+            .post(`/v1/webhooksv1`)
             .set('cookie', `${SESSION_COOKIE_NAME}_1=${cookie}`)
             .set('x-fp-signature', getSignature(reqBody))
             .send(reqBody);