+++ types/express/retry_manager.js
@@ -1,0 +1,4 @@
+const { RetryManger } = require('../retry_manager');
+module.exports = {
+    RetryManger
+};