+++ types/lib/constants.js
@@ -1,0 +1,12 @@
+'use strict';
+exports.SESSION_COOKIE_NAME = "ext_session";
+exports.ADMIN_SESSION_COOKIE_NAME = 'ext_adm_session';
+exports.TEST_WEBHOOK_EVENT_NAME = "ping";
+exports.ASSOCIATION_CRITERIA = {
+    ALL: "ALL",
+    SPECIFIC: "SPECIFIC-EVENTS",
+    EMPTY: "EMPTY"
+};
+exports.TIMEOUT_STATUS = 504;
+exports.SERVICE_UNAVAILABLE = 503;
+exports.BAD_GATEWAY = 502;