+++ types/lib/extension.js
@@ -6,9 +6,8 @@ const { PlatformConfig, PlatformClient, PartnerConfig, PartnerClient } = require
 const { WebhookRegistry } = require('./webhook');
 const logger = require('./logger');
 const { fdkAxios } = require('@gofynd/fdk-client-javascript/sdk/common/AxiosHelper');
-const { version } = require('./../package.json');
-const { RetryManger } = require("./retry_manager")
-
+const { version } = require('../package.json');
+const { RetryManger } = require("./retry_manager");
 class Extension {
     constructor() {
         this.api_key = null;
@@ -23,35 +22,26 @@ class Extension {
         this._retryManager = new RetryManger();
         this.configData = null;
     }
-
     async initialize(data) {
-
         if (this._isInitialized) {
             return;
         }
-
         this._isInitialized = false;
         this.configData = data;
-
         this.storage = data.storage;
-
         if (!data.api_key) {
             throw new FdkInvalidExtensionConfig("Invalid api_key");
         }
         this.api_key = data.api_key;
-
         if (!data.api_secret) {
             throw new FdkInvalidExtensionConfig("Invalid api_secret");
         }
         this.api_secret = data.api_secret;
-
         if (!data.callbacks || (data.callbacks && (!data.callbacks.auth || !data.callbacks.uninstall))) {
             throw new FdkInvalidExtensionConfig("Missing some of callbacks. Please add all `auth` and `uninstall` callbacks.");
         }
-
         this.callbacks = data.callbacks;
         this.access_mode = data.access_mode || "offline";
-
         if (data.cluster) {
             if (!validator.isURL(data.cluster)) {
                 throw new FdkInvalidExtensionConfig("Invalid cluster value. Invalid value: " + data.cluster);
@@ -61,9 +51,7 @@ class Extension {
         else
             data.cluster = this.cluster;
         this.webhookRegistry = new WebhookRegistry(this._retryManager);
-
         await this.getExtensionDetails();
-
         if (data.base_url && !validator.isURL(data.base_url)) {
             throw new FdkInvalidExtensionConfig("Invalid base_url value. Invalid value: " + data.base_url);
         }
@@ -71,25 +59,19 @@ class Extension {
             data.base_url = this.extensionData.base_url;
         }
         this.base_url = data.base_url;
-
         if (data.scopes) {
             data.scopes = this.verifyScopes(data.scopes, this.extensionData);
         }
         this.scopes = data.scopes || this.extensionData.scope;
-
         logger.debug(`Extension initialized`);
-
         if (data.webhook_config && Object.keys(data.webhook_config)) {
             await this.webhookRegistry.initialize(data.webhook_config, data);
         }
-
         this._isInitialized = true;
     }
-
     get isInitialized() {
         return this._isInitialized;
     }
-
     verifyScopes(scopes, extensionData) {
         const missingScopes = scopes.filter(val => extensionData.scope.indexOf(val) === -1);
         if (!scopes || scopes.length <= 0 || missingScopes.length) {
@@ -97,17 +79,14 @@ class Extension {
         }
         return scopes;
     }
-
     getAuthCallback() {
         return urljoin(this.base_url, "/fp/auth");
     }
-
     isOnlineAccessMode() {
         return this.access_mode === 'online';
     }
-
     async getPlatformConfig(companyId) {
-        if (!this._isInitialized){
+        if (!this._isInitialized) {
             await this.initialize(this.configData);
         }
         let platformConfig = new PlatformConfig({
@@ -116,24 +95,20 @@ class Extension {
             apiKey: this.api_key,
             apiSecret: this.api_secret,
             useAutoRenewTimer: false,
-            logLevel: this.configData.debug === true? "debug": null
+            logLevel: this.configData.debug === true ? "debug" : null
         });
         return platformConfig;
-
     }
-
     async getPlatformClient(companyId, session) {
-        if (!this._isInitialized){
+        if (!this._isInitialized) {
             await this.initialize(this.configData);
         }
         const SessionStorage = require('./session/session_storage');
-        
         let platformConfig = await this.getPlatformConfig(companyId);
         platformConfig.oauthClient.setToken(session);
         platformConfig.oauthClient.token_expires_at = session.access_token_validity;
-
         if (!session.access_token_validity || session.refresh_token) {
-            let ac_nr_expired = !session.access_token_validity? true: ((session.access_token_validity - new Date().getTime()) / 1000) <= 120;
+            let ac_nr_expired = !session.access_token_validity ? true : ((session.access_token_validity - new Date().getTime()) / 1000) <= 120;
             if (ac_nr_expired) {
                 logger.debug(`Renewing access token for company ${companyId} with platform config ${logger.safeStringify(platformConfig)}`);
                 const renewTokenRes = await platformConfig.oauthClient.renewAccessToken(session.access_mode === 'offline');
@@ -146,36 +121,31 @@ class Extension {
         let platformClient = new PlatformClient(platformConfig);
         platformClient.setExtraHeaders({
             'x-ext-lib-version': `js/${version}`
-        })
+        });
         return platformClient;
     }
-
     getPartnerConfig(organizationId) {
         if (!this._isInitialized) {
             throw new FdkInvalidExtensionConfig("Extension not initialized due to invalid data");
         }
-
         let partnerConfig = new PartnerConfig({
             organizationId: organizationId,
             domain: this.cluster,
             apiKey: this.api_key,
             apiSecret: this.api_secret,
             useAutoRenewTimer: false,
-            logLevel: this.configData.debug ===  true? "debug": null
-        })
+            logLevel: this.configData.debug === true ? "debug" : null
+        });
         return partnerConfig;
     }
-
     async getPartnerClient(organizationId, session) {
         if (!this._isInitialized) {
-            throw new FdkInvalidExtensionConfig('Extension not initialized due to invalid data')
+            throw new FdkInvalidExtensionConfig('Extension not initialized due to invalid data');
         }
         const SessionStorage = require('./session/session_storage');
-
         let partnerConfig = this.getPartnerConfig(organizationId);
         partnerConfig.oauthClient.setToken(session);
         partnerConfig.oauthClient.token_expires_at = session.access_token_validity;
-
         if (session.access_token_validity && session.refresh_token) {
             let ac_nr_expired = ((session.access_token_validity - new Date().getTime()) / 1000) <= 120;
             if (ac_nr_expired) {
@@ -190,25 +160,18 @@ class Extension {
         let partnerClient = new PartnerClient(partnerConfig);
         partnerClient.setExtraHeaders({
             'x-ext-lib-version': `js/${version}`
-        })
+        });
         return partnerClient;
     }
-
     async getExtensionDetails() {
-
         let url = `${this.cluster}/service/panel/partners/v1.0/extensions/details/${this.api_key}`;
         const uniqueKey = `${url}`;
-
         const retryInfo = this._retryManager.retryInfoMap.get(uniqueKey);
         if (retryInfo && !retryInfo.isRetry) {
             this._retryManager.resetRetryState(uniqueKey);
         }
-
         try {
-            const token = Buffer.from(
-                `${this.api_key}:${this.api_secret}`,
-                "utf8"
-            ).toString("base64");
+            const token = Buffer.from(`${this.api_key}:${this.api_secret}`, "utf8").toString("base64");
             const rawRequest = {
                 method: "GET",
                 url: url,
@@ -221,24 +184,18 @@ class Extension {
             let extensionData = await fdkAxios.request(rawRequest);
             logger.debug(`Extension details received: ${logger.safeStringify(extensionData)}`);
             this.extensionData = extensionData;
-        } catch (err) {
-
-            if (
-                RetryManger.shouldRetryOnError(err) 
-                && !this._retryManager.isRetryInProgress(uniqueKey)
-            ) { 
-                logger.debug(`API call failed. Starting retry for ${uniqueKey}`)
+        }
+        catch (err) {
+            if (RetryManger.shouldRetryOnError(err)
+                && !this._retryManager.isRetryInProgress(uniqueKey)) {
+                logger.debug(`API call failed. Starting retry for ${uniqueKey}`);
                 return await this._retryManager.retry(uniqueKey, this.getExtensionDetails.bind(this));
             }
-
             throw new FdkInvalidExtensionConfig("Invalid api_key or api_secret. Reason:" + err.message);
         }
     }
 }
-
-
 const extension = new Extension();
-
 module.exports = {
     extension
 };